requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
motor
pymongo
openai
openai-whisper
torch
langchain
langchain-openai
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
boto3
fastapi-users
networkx
pytest
pytest-asyncio
email-validator
